Adds support for installing Chef client on Windows
This includes splitting the installer into platform-specific
implementations.

Due to an interface mismatch between the win_rm and winrm_fs gems
file_upload does not work over winrm at this time; so the
windows installation will be forced to do remote-download
of chef client.

This will need to be rebased after SHACK-86 lands

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>