Slight cleanups to error handling
This consolidates the interface behind a singleton method so that
callers don't have to do an extra 'unwrap' step prior to invoking.

This also adds a common dump_unexpected_error class method to ensure
that we always show an error, even when things go wrong enough
that our error handling itself fails.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>