Bug fixes
This fixes two issues:

First, during demos I discovered that if the remote command failed but did so before chef was invoked, we could end up getting a stale chef stacktrace copied over. So now we delete the chef stacktrace after we copy it to the workstation - this prevents copying over an old, stale stacktrace and hindering bug troubleshooting.

Second, we were trying to output some terminal information before initializing the terminal (in the scenario where a config file has not yet been created).