Make plain-text spinner element avaialble via config
This adds a PlainTextElement that has an interface like a TTY::Spinner
but provides update-per-line plain text output.

This is useful when debugging with pry while a spinner
is running; without it, the spinner refreshes interfere
with the IRB session.

To enable this, add the following to config.toml:

[dev]
spinner="PlainTextElement"

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>