Make action implmenetations unaware of UI concerns
Updates Action::Base.run to accept a block.  The block is invoked
whenever the action wishes to notify the caller that a step has
completed.

Implementations for InstallAction and TargetConvergeAction have been
updated to no longer reference UI text, but instead notify
on events.  The calling command now manages updating text based on the
event.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>