Show help in a consistent manner
Ensure help works consistently.

If help is visible anywhere in the command as --help, -h, or help
this will treat it as the 'help' flag and ensure that the lowest
level subcommand (before the 'help' word was encountered) is the one
that renders help.

Version has been moved to its own command, but we have retained a '-v'
flag for easy-of-use.