Add handling for multi-target errors
This wraps the actions for a multi-target command
(converge in this case) in a Terminal::Job instance,
which tracks its failure state.

After all jobs are complete, this collects those that failed
and sends them over to ErrorPrinter for formatting and capture
to file, while a top level MULTI001 (basically: one or more jobs failed)
error is shown to the user with the location of the errors file.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>