include a more complete collection of tooling from DK
Copying the software definitions from Chef DK and the fiddly bits they
depend on so that Workstation can install all the tooling that DK
installs. Copying code is bad, m'kay, but seemed to be the simplest for
the moment to (1) avoid needing to move the software definitions themselves
in the DK repo (to config/software/) or (2) changing omnibus to include
omnibus/config/software/ in definition search paths.

Signed-off-by: Robb Kidd <robb@thekidds.org>