[SHACK-129] Use policyfile for remote code bundle
This changes a few things along the way. First, we make a temporary
cookbook locally, whatever the source (local recipe file or command line
resource) is. Then we bundle that local cookbook into a policyfile
bundle. Finally we ship that bundle to the remote machine and execute
chef-client on it, pointing chef directly at the bundle instead of
having to extract it first.

This functionality requires chef client >= 14.1.1 so we are now
installing from the unstable channel.