Add --no-install flag to bypass install/upgrade
The --no-install flag will prevent install or upgrade of the target's
chef client installation. If an update or installation is required,
an appropriate error message is shown with instructions on permitting
the change.

This also improves in-task error handling so that the
UI gets a chance to be refreshed when a task fails. Previously
the error would be reported, but the summary 'spinner' line would
never be updated for the failure.