pin train to pre-GCP version
Train 1.4.7 added Google GCP libraries. The google-protobuf gem does not coexist
in an omnibus build well because its C-extensions build seemingly for
every version of Ruby which are then caught by the omnibus health check for
.so files built against non-existent libraries.

Example of failure can be found on https://github.com/google/protobuf/issues/4210

Applying this pin WHICH SHOULD TOTALLY BE REMOVED when we have solved
for omnibus packaging of google-protobuf.

Signed-off-by: Robb Kidd <robb@thekidds.org>