Give embedded username preference over --user flag value
By default, train will prioritize the :user flag higher than
the username value included in a host url, eg ssh://user1@host.
This makes less sense in a tool that allows multiple targets to be
specified, with some that have different username requirements than
others.

This changes TargeResolver so that the embedded name takes priority,
and the default is used when no embedded name is present.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>