resolve regression on telemetry not actually sending data
* Symptoms

YAML files pile up in `~/.chef-workstation/telemetry`.

* Signs

No telemetry data from a `chef-run` appears at an endpoint.

`ChefRun::CLI.run()` was not able to instantiate a telemeter sender:

     => 120:       Thread.new() { ChefRun::Telemeter::Sender.new().run }

    [1] pry(#<ChefRun::CLI>)> Thread.new() { ChefRun::Telemeter::Sender.new().run }
    => #<Thread:0x00007fc8a2b8d310@(pry):1 dead>
    [2] pry(#<ChefRun::CLI>)> ChefRun::Telemeter::Sender.new().run
    NameError: uninitialized constant ChefRun::Telemeter::Sender
    from (pry):2:in `run'

The uninitialized constant exception was lost in the thread spun off for
sending telemetry data, so the attempted send amounted to a no-op.

* Treatment

Added the necessary require statements for `ChefRun::CLI` to know about
the telemetry data sender class for it to Do The Thing™.

Signed-off-by: Robb Kidd <robb@thekidds.org>