Make error handling within connect_target correct
The logic for not having a reporter should not have been there,
since we always now have a reporter; but this bug meant we were hitting
that nil path and not correctly handling the error reporting.

This change splits it into a common function with error handling, and
invokes it differently based on single/multi target execution.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>