Ensure that all errors bubble up
The 'ensure' clause in 'run' was executing as it should - but
when it ran after an exception, the exception was not raising. This was
causing us to get no output at all when errors (like LoadError) occurred
where we didn't expect them.

I created handle_run_error which returns the error code, and we now use
it both in the `timed_run_capture` and in `run` itself in event of
exception.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>