Revert train to 1.4.6, update sudo error checks
This reverts train to 1.4.6.  Later versions introduce a gcp
dependency that does not cleanly pass omnibus health checks.  While that
is being resolved, we'll continue to stick to 1.4.6.

We were using the Train::Error#reason field added in 1.4.12; this PR
updates error handling to do string-based matching of the messages
provided by train instead.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>