Always expand host to include full user info
To provide a consistent experience with multi- and single-target,
we will always construct a fully qualified url in the form:

proto://user-info@target

Where user-info is "username:password" (ref: RFC 3928[1] s. 3.2.1)

We no longer pass the user or password option into train, so that there
is no conflict between those values and the values embedded in the url.
This ensures that train will only have one set of credentials to work
with per host, and won't do surprising things.

A blank username will be expanded in the form ":password@host.com" and
a blank password will be supplied in the form "username:@host.com", both
of which are valid under the RFC.

This fixes github issue #175
[1] https://www.ietf.org/rfc/rfc3986.txt
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>Always expand host to include full user info
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>