Fix error that occurs when config file deleted.
When the default ocnfiguration file is deleted, and
a custom file is not specified, chef-run emits a stack trace
due to failure loading the missing file.

This can happen when the customer deletes or removes the default
config.toml. chef-run creates the file only on first-run, so if it's
removed after that point it stays gone.

Because we have to assume the customer had a reason for deleting the
configuration file, we can't recreate a new empty file.  Instead,
this fix changes it so that if the config file is missing, we don't try
to load it. This will leave configuration defaults in place.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>