Create ~/.chef-workstation before config file
Attempting to run these programs without ~/.chef-workstation already
created results in a Ruby stack trace as it fails to write a config file
to that path.

This change adds the requisite mkdir commands to make sure the directory
exists before attempting to write under it. Note--I didn't copy the
"unless exist" clause as used in the integration tests because mkdir is
idempotent and I trust it's actually less expensive without our
condition.

Signed-off-by: Trevor Bramble <inbox@trevorbramble.com>