Our verify scripts should fail the pipeline if any of our commands fail
Signed-off-by: tyler-ball <tball@chef.io>

Update the exit handling output for the verify script. It's worth
noting, because it's really hard to find info about this, that
using 'call' effectively blocks on that command, which is nice.
The other thing is that ERRORLEVEL in the context of a single cmd
execution will not be reset unless one specifically runs 'call set'
in the script, so in this case the single ERRORLEVEL line acts
as a rollup of any errors (the last failure code is saved),
which works great for us. Finally, :error is actually a 'magic'
keyword that refers to the error handler of the script that
actually invokes this verify script.

Signed-off-by: Scott Hain <shain@chef.io>

Update: I did not find that ERRORLEVEL was a rollup during testing.
I had to put it after each command execution - not sure, but got
something that works

Signed-off-by: tyler-ball <tball@chef.io>