Show more ssh connection error detail
In order to present a meaningful message detailing
the nature of the auth problem when ssh auth fails,
unwrap the Train::Error to use '.cause' when it's present.

We still have some improvements to make here, but this will
let us tell the user what went wrong under the 'cheftrn999' error
code instead of a simple 'connection failed' message.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>