[SHACK-304] Subscribe to ChefDK merges in expeditor
This means that every time a ChefDK PR is merged we will get a new build
of Chef Workstation. This new Chef Workstation build pulls ChefDK from
master for the omnibus build so effectively means a we will get a new
Chef Workstation build for every minor version bump (PR merge) of
ChefDK.

This could potentially include a build of ChefDK that does not pass the
pipeline since the ChefDK pipeline build occurs asynchronously from the
expeditor merge action.

We also update the omnibus definition to only include ChefDK since that
pulls Chef Apply in as a gem dependency now.

Signed-off-by: tyler-ball <tball@chef.io>