Add chef-workstation-tray app to chef-workstation
For the short term, this software definition grabs the current most
release for the build platfrom from the tray app github releases.

The behavior differs by platform:

Linux:
 - download the AppImage file, install it as executable
 '/opt/chef-workstation/bin/chef-tray'
Windows:
- download the MSI file to /opt/chef-workstation/chef-tray.msi
Mac:
- download the MSI file to /opt/chef-workstation/chef-tray.dmg

In this iteration, Windows and Mac operators will need to explicitly
install the tray app from the included installer file. Future updates
are  expected to make this automatic.

Earlier iterations attempted to build the package locally, but we ran
into various platform-specific issues on different target platforms which led us
to this point.

Because this is only an experimental path, we switched
to the current option so that we could sooner explore how we could manage handling the actual
installation of the tray app once the package is on the machine.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>