Install platform-specific package on linux
This replaces appimage.

The reason for this is that continuing to use appimage as we were would
prevent a simple upgrade when a new tray app is avaialble - the user
would have to rename the download and replace
`chef-workstation/bin/chef-tray` with it.

Using the platform-specific package allows the installation to be
be managed using familiar tools (ie dpkg or rpm), and avoids changing
content of the chef-workstation omnibus installation which can cause problems
when it's time to uninstall/upgrade the omnibus package.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>