Run npm from the build dir instead of installing
Instead of temporarily installing nodejs into /opt/chef-workstation
this keeps it in place in its `project_dir`; and chef-workstation-app
uses it from that location to package the electron app.

This resolves issues on builders where cleanup of install runs,
then a new build runs and the nodejs bins are not present because
nodejs-binary is cached and doesn't need to be rebuilt.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>