Make CWA available on Linux.
This updates the chef-workstation-app software definition to whitelist node-related binaries,
and do directly deploy the CWA files to
$install_dir/components/chef-workstation-app.

In postinst, it adds a simple check to see if CWA is runnable by verifying required libs are present.
If it is, it will link it into $install/bin. Otherwise, it prints an informational
message letting the operator know that chef-workstation-app will not be available.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>