Add Chef Workstation App installer feature for Windows
This splits the Chef Workstation windows installer into
'Chef Workstation' and 'Experimental Features'.

'Chef Workstation' contains the existing items for this package,
while 'Experimental Features' currently provides the option to
install Chef Workstation Tray.  Selecting this will install
autostart, desktop, and Start shortcuts for it.

This is intended to be a short-term approach while we work on
moving these types of feature flags into configuration and the
Workstation App itself.

Additional changes:

* renames the shell shortcut "Chef Workstation" to "Chef
  Workstation Command Prompt". This is more in line with recent
  discussions around naming, and avoids confusion between a shortcut
  called "Chef Workstation" and one named "Chef Workstation App".
* adds clarifying wording to Windows installation components and
  descriptions now that there can be more than one 'Chef Workstation' icon
* shorten key names

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>