[SHACK-352] Enable CWA on Mac OS X
This packages the Mac version of the chef-workstation-app in the mac
build as a zip file.  At package install time, it will add Chef
Workstation App to /Applications by extracting it from the zip file
and moving it into place.

This differs slightly from what we do on linux and windows (we include
the unpacked CWA in those cases); this is because Mac's pkgbuild will
recognize the contained "Chef Workstation App.app" directory as a
deployable application during the omnibus build , and treat it
differently - making it unavailable for us to set up as part of the
package install.

Signed-off-by: Tyler Ball <tball@chef.io>
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>