Begin using release rollup in CHANGELOG
This collects all changes since the last `stable` promotion in a section
at the top of the CHANGELOG. This is a good way to display a snapshot of
the changes that exist in the latest `current` build to end users...say
by maybe a menu tray applicaiton. :-)

I also took the time to properly format the CHANGELOG for all existing
`stable` releases.

Signed-off-by: Seth Chisamore <schisamo@chef.io>