Pin delivery-cli version and add subscription
Now that A2 includes workflow there may be features added to delivery-cli.
In the past we always just pulled master whenever we built workstation, but
this means we only pick up delivery-cli changes when we build for other reasons.
This change pins the version and adds an expeditor subcription that updates a
new version override in our omnibus project definition so we can stay up
to date.

Signed-off-by: Jon Morrow <jmorrow@chef.io>