Two builds, one repository: chefdk build via WS
Experimental support for building both chef-dk
and chef-workstation.  This commit adds a chefdk project to
chef-workstation's omnibus configuration, and brings
chef-workstation's software definitions up to date with chef-dk.

This does not address pipeline changes - it only provides a means to
build both packages from the chef-workstation repository.

I've validated this so for DK and WS omnibus build, eg
`omnibus build chefdk` and `omnibus build chef-workstation`.

The former works and generates a DK package. The latter fails in the
electron setup for chef-workstation-app - I don't think this is related
but will dig further.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>