Add the 'gems' component.
The `gems` component is used by appbundler
to build all of the gems we ship as part of
Chef Workstation.

The Gemfile.lock in this directory is the One True Source
of shipped gems.

It is referenced from the  gems software definition in
omnibus/config/software/gems.rb.

This Gemfile and the update_gemfile_lock task was pulled from
ChefDK's as of v4.0.74, and will be maintained independently
going forward.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>