Exclude development mode in TK and berkshelf
After removing our dependency on the cucumber gem,
the development versions of cucumber in TK and berkshelf
were causing cascading appbundler dep failures.

It seems to have caused TK and berkshelf to pull in a
version of `cucumber` that was incompatible with the `cucumber-core`
required by foodcritic (not in dev mode).

`--without development` for TK and berkshelf prevents
the full version of `cucumber` from getting pulled in at all.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>