Use --without when appbundling
We were pulling in more gems than we needed by not excluding
the right groups in appbundler.

This is also required given recent changes in which we
stopped shipping cucumber - testkitchen will not appbundle correctly
without it unless 'test' is excluded.

This commit also groups the binary wrapping with the
gems the binaries are coming from, for the sanity of those who come
after.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>