Add libGlesV2 to whitelist chef-workstation-app.
We recently upgraded to electron v6 [1], which introduced this new dependency.

Because we have not unpacked the electron build dependencies for
inclusion into omnibus, we're adding this to the whitelist.
Similar to the other whitelisted components, this means that
chef-workstation-app will not run when they are missing on the target
Linux system.

We currently have a check in postinstall.sh that warns when this
happens.

[1] https://github.com/chef/chef-workstation-app/commit/33b278d4468a0095fea050d280ca49f1a688a06c
pour

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>