Manage package verification as part of Workstation
Package verification in CI is currently performed by running the
hidden `chef-cli` command `chef verify`. This command references all of
the tools packaged with Workstation, but is not a part of Workstation
and does not depend on it.

This change moves the verification steps into a bash/ps1 script
that is executed in the CI pipeline against the built omnibus packages.

Signed-off-by: Your Name <you@example.com>