Transplant verify functionality from chef-cli
This moves the verify command into chef-workstation, since
it is only used to verify the builds in the omnibus pipeline.
This removes the indirect circular dependency between
Workstation and CLI as well, since CLI no longer needs to know
about what tools ship with Workstation.

This is a minimally modified transplant of the Verify class and
the supporting class ComponentTest from chef-cli. Once all is well,
we can update chef-cli to remove the verify command.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>