Pass only arguments to sub-binaries/add debug msgs
This commit is detecting properly when a sub-binary was triggered
successfully and it returned an exit code different than 0 and pass it
directly to the main wrapper for proper propagation.

This commit also passes only the arguments to the sub-binaries as well
as a better way to handle debug messages enabled with the environment
variable `DEBUG`. Example:
```
$ DEBUG=true chef knife cookbook list
Bin: /usr/local/bin/knife
Args: [cookbook list]
foo   0.1.0
```

Signed-off-by: Salim Afiune <afiune@chef.io>