Make a `chef` binary and use it
Trying to symlink `chef` to `chef-cli` causes difficulty
on windows in particular, because we don't symlink on that platform and
the `chef` binary does not exist.

Instead, we'll make a copy of chef-cli as `chef` in the install bin dir,
and symlink it like we do the other commands (eg /usr/bin/chef ->
/opt/chef-workstation/bin/chef).

When we replace the chef command, we won't need to modify
postinst/postrm any further - they'll continue symlinking `chef` and
removing the symlink, respectively.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>