Update the list of components we ship
This updates the readme and installer list of components that we ship to
match what we display in the chef -v output + a mention of knife/kitchen
plugins. This gives users a better idea of what we're shipping in this
package. I also removed all the remaining mentions to DK I could find in
the codebase.

Signed-off-by: Tim Smith <tsmith@chef.io>