Hide chef-analyze behind a feature flag
Since we are developing new tools and new features, we would like to be
able to ship them fast to iterate over them, also to have feedback from
users that wants to test experimental functionalities.

This change is hiding the new `chef-analyze` behind a feature flag.

To enable this feature use the environment varible:
```
CHEF_FEAT_ANALYZE=true chef analyze
```

Or, to enable this feature permanently, modify your `$HOME/.chef-workstation/config.toml` with:
```
[features]
analyze = true
```

When the feature is not enabled, the user will see a message that
explains how to enable the experimental functionality:
```
➜  ~ chef analyze

This fuctionality is still under development and is cosidered experimental.

You can enable this feature through the environment varible:

CHEF_FEAT_ANALYZE=true

Or, to enable this feature permanently, modify your $HOME/.chef-workstation/config.toml with:

[features]
analyze = true

```

Signed-off-by: Salim Afiune <afiune@chef.io>