[bug-fix] Enable interactive 'chef' commands
Currently, any `chef` command that requires STDIN interaction is broken,
things like `chef generate cookbook` will ask for accepting the license
and we are unable to do so. Another example is `chef exec irb`.

This PR fixes that problem by linking the STDIN from the terminal to the
underlying spawned command.

Closes https://github.com/chef/chef-workstation/issues/637

Signed-off-by: Salim Afiune <afiune@chef.io>