Update Cookstyle to 6.0 / Chefstyle to 0.15.1
This adds 7 new cops, adds autocorrection to a good number of cops, improves offense detection, and enables new built-in Rubocop cops. It also resolves several failures I reported in the Rubocop engine that caused cookstyle to fail when checking cookbooks on the supermarket. Chefstyle is bumped as well since the Rubocop engine versions need to match.

Signed-off-by: Tim Smith <tsmith@chef.io>