Manually building and installing docker-api gem
The ruby cleanup omnibus software dep removes the gem installed by
bundler. I am not sure how Ohai does not have this issue. Reverting
to a pattern I understand to fix our build for now.

In the future it would be nice to only have to manage this gem via
the Gemfile and without the docker-api omnibus software definition.

Signed-off-by: tyler-ball <tball@chef.io>