Note that users can upload to an existing server if they maintain their pins
This addresses valid concerns that not all customers will be able to
create a new Chef Server and bootstrap nodes to that. This guide
currently combines two tasks (upgrading client version, and migrating to
a new server) but we feel that doing this creates a clean slate to
build from in the future.

Signed-off-by: tyler-ball <tball@chef.io>