Simplify message shown when WS app missing deps
When the Workstation App is missing dependencies on linux,
the postinst script spams the screen with a long list of missing
libraries.

This PR modifies the messaging to exclude the missing dependency list, and
recommend installing the 'electron' app to meet them.

This is not ideal, but given the constraints of omnibus (optional
package dependencies are not supported) and our desire to ensure that
the desktop application is included in the Workstation package,
it seems a reasonable compromise.

In the future, a cleaner path would be to build electron (and its deps) in the
omnibus build of Workstation.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>