Bump cookstyle and treetop to resolve 2 issues
Cookstyle was no longer detecting bad ruby syntax due to a change (fix?)
in the RuboCop engine. This is one of the most useful parts of cookstyle
so this is a pretty major regression

See https://github.com/chef/cookstyle/issues/733

Treetop was monkeypatching String.blank? which was breaking the usage of
Ruby's implementation by Cookstyle/RuboCop. This was causing:

https://github.com/rubocop-hq/rubocop/issues/8636

Signed-off-by: Tim Smith <tsmith@chef.io>