WS powershell on windows does not require runas
This changes the verb used to launch powershell and cmd on
Windows from 'runas' to 'open'.  This reduces the permissions required,
so that operators are no longer shown the UAC/admin prompt
that displays when starting WS powershell on Windows.

Note that the UAC prompt on Windows concealed a slow startup
time for the WS powershell, because it uses `chef shell-init` to
configure the environment.  This will be addressed separately.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>